Receive full 64-bit NTP timestamp
Add the facility to extract the fractional part of the NTP response,
allowing the recieved time to be exposed as a canonical NTP timestamp,
comprising a 32-bit seconds field and a 32-bit fractional field. This
allows the returned time value of a higher precision than was previously
available.

Convert the functions returning raw timestamps to instead return a
struct encapsulating an NTP timestamp.

Also add an example for raw timestamp reception.